add options for path to local source code and define language manually in ToolDog option
